add carousel to components
