+++ src/components/carousel/styles.scss
@@ -45,7 +45,7 @@
     display: none;
     display: inline-block;
     position: absolute;
-    top: calc(50% - 30px);
+    top: calc(50% - 20px);
     z-index: 10;
 
     &.slick-disabled {
